+++ resources/views/admin/phrase/list.blade.php
@@ -4,10 +4,21 @@
     @if (Session::has('message'))
         <p class="alert alert-danger">{!! Session::get('message') !!}</p>
     @endif
+    <div class="panel-body">
+        <a href="{{ url('listphrases') }}" data-toggle="modal" class="btn btn-success disabled" disabled>
+            Listado de frases Aprobadas
+        </a>
+        <a href="{{ url('pending') }}" data-toggle="modal" class="btn btn-warning">
+            Listado de frases Pendientes a Aprobar
+        </a>
+        <a href="{{ url('reject') }}" data-toggle="modal" class="btn btn-danger">
+            Listado de frases Rechazadas
+        </a>
+    </div>
     <div class="table-agile-info">
         <div class="panel panel-default">
             <div class="panel-heading">
-                Frases
+                Frases Aprobadas
             </div>
             <div class="table-responsive">
                 <table class="table table-striped b-t b-light">
@@ -26,16 +37,25 @@
                             <td>{{ $phrase->text }}</td>
                             <td>{{ $phrase->author }}</td>
                             <td>{{ $phrase->category->category }}</td>
-                            <td>{{ $phrase->user_id }}</td>
-                            <td>
-                                <a href="{{ url('phrase', $phrase->id) }}" type="submit" class="btn btn-link"><i
-                                            class="fa fa-check text-success text-active"></i></a>
-                                {!! Form::open(['method' => 'DELETE','route' => ['phrase.destroy', $phrase->id],'style'=>'display:inline']) !!}
-                                {{Form::token() }}
-                                <button type="submit" class="btn btn-link" >
-                                    <i class="fa fa-times text-danger text"></i></button>
-                                {!! Form::Close() !!}
-                            </td>
+                            <td>{{ $phrase->User->name }}</td>
+                            @if(Auth::user()->type == 'ADMIN')
+                                <td>
+                                    <a href="{{ url('phrase', $phrase->id) }}" type="submit" class="btn btn-link"><i
+                                                class="fa fa-pencil text-success text-active"></i></a>
+                                    {!! Form::open(['method' => 'DELETE','route' => ['phrase.destroy', $phrase->id],'style'=>'display:inline']) !!}
+                                    {{Form::token() }}
+                                    <button type="submit" class="btn btn-link">
+                                        <i class="fa fa-trash text-danger text"></i></button>
+                                    {!! Form::Close() !!}
+                                </td>
+                            @else
+                                <td>
+                                    <button type="submit" class="btn btn-link disabled"><i
+                                                class="fa fa-pencil text-success text-active"></i></button>
+                                    <button type="submit" class="btn btn-link disabled">
+                                        <i class="fa fa-trash text-danger text"></i></button>
+                                </td>
+                            @endif
                         </tr>
                     @endforeach
                     </tbody>
